Fixed warning in 64bit builds
This fixes

tracegen.cpp(48) : warning C4244: 'return' : conversion from 'WPARAM'
                   to 'int', possible loss of data

+++ tracegen.cpp
@@ -45,5 +45,5 @@ int WINAPI WinMain( HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLin
         }
     }
 
-    return msg.wParam;
+    return (int)msg.wParam;
 }
